it surprisingly doesn't trigger page fault anymore It works neither.

+++ include/lib/libc/string.h
@@ -5,7 +5,7 @@
 
 extern "C"
 {
-    void *memset(void *, size_t, size_t);
+    void *memset(void *, int, size_t);
     int memcmp(const void *, const void *, size_t);
     void *memmove(void *, const void *, size_t);
     int strlen(const char *);
